+++ src/Products/PloneMeeting/MeetingConfig.py
@@ -1338,23 +1338,6 @@ schema = Schema((
         write_permission="PloneMeeting: Write risky config",
     ),
     LinesField(
-        name='meetingColumns',
-        widget=MultiSelectionWidget(
-            description="MeetingColumns",
-            description_msgid="meeting_columns_descr",
-            format="checkbox",
-            label='Meetingcolumns',
-            label_msgid='PloneMeeting_label_meetingColumns',
-            i18n_domain='PloneMeeting',
-        ),
-        schemata="gui",
-        multiValued=1,
-        vocabulary='listMeetingColumns',
-        default=defValues.meetingColumns,
-        enforceVocabulary=True,
-        write_permission="PloneMeeting: Write risky config",
-    ),
-    LinesField(
         name='availableItemsListVisibleColumns',
         widget=MultiSelectionWidget(
             description="availableItemsListVisibleColumns",
@@ -1385,6 +1368,40 @@ schema = Schema((
         multiValued=1,
         vocabulary='listItemsListVisibleColumns',
         default=defValues.itemsListVisibleColumns,
+        enforceVocabulary=True,
+        write_permission="PloneMeeting: Write risky config",
+    ),
+    LinesField(
+        name='itemActionsColumnConfig',
+        default=defValues.itemActionsColumnConfig,
+        widget=MultiSelectionWidget(
+            description="ItemActionsColumnConfig",
+            description_msgid="item_actions_column_config_descr",
+            format="checkbox",
+            label='Itemactionscolumnconfig',
+            label_msgid='PloneMeeting_label_itemActionsColumnConfig',
+            i18n_domain='PloneMeeting',
+        ),
+        schemata="gui",
+        multiValued=1,
+        vocabulary='listItemActionsColumnConfig',
+        enforceVocabulary=True,
+        write_permission="PloneMeeting: Write risky config",
+    ),
+    LinesField(
+        name='meetingColumns',
+        widget=MultiSelectionWidget(
+            description="MeetingColumns",
+            description_msgid="meeting_columns_descr",
+            format="checkbox",
+            label='Meetingcolumns',
+            label_msgid='PloneMeeting_label_meetingColumns',
+            i18n_domain='PloneMeeting',
+        ),
+        schemata="gui",
+        multiValued=1,
+        vocabulary='listMeetingColumns',
+        default=defValues.meetingColumns,
         enforceVocabulary=True,
         write_permission="PloneMeeting: Write risky config",
     ),
@@ -4421,6 +4438,22 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         """ """
         return []
 
+    security.declarePrivate('listItemActionsColumnConfig')
+
+    def listItemActionsColumnConfig(self):
+        d = "PloneMeeting"
+        res = DisplayList(())
+        for prefix, translatable_value in (('', ''),
+                                           ('meetingmanager_', ' (MeetingManager)'),
+                                           ('manager_', ' (Manager)')):
+            res.add(prefix + "delete",
+                    translate('Item action delete' + translatable_value, domain=d, context=self.REQUEST))
+            res.add(prefix + "duplicate",
+                    translate('Item action duplicate' + translatable_value, domain=d, context=self.REQUEST))
+            res.add(prefix + "history",
+                    translate('Item action history' + translatable_value, domain=d, context=self.REQUEST))
+        return res
+
     security.declarePrivate('listVotesEncoders')
 
     def listVotesEncoders(self):
@@ -5042,7 +5075,7 @@ class MeetingConfig(OrderedBaseFolder, BrowserDefaultMixin):
         for mctct in self.getMeetingConfigsToCloneTo():
             configId = mctct['meeting_config']
             actionId = self._getCloneToOtherMCActionId(configId, self.getId())
-            urlExpr = "string:javascript:event.preventDefault();callViewAndReload(base_url='${object_url}', " \
+            urlExpr = "string:javascript:callViewAndReload(base_url='${object_url}', " \
                 "view_name='doCloneToOtherMeetingConfig', " \
                 "params={'destMeetingConfigId': '%s'});" % configId
             availExpr = 'python: object.meta_type == "MeetingItem" and ' \