+++ src/Products/PloneMeeting/MeetingItem.py
@@ -5540,7 +5540,8 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
               cloneEventAction=None, cloneEventActionLabel=None, destFolder=None,
               copyFields=DEFAULT_COPIED_FIELDS, newPortalType=None, keepProposingGroup=False,
               setCurrentAsPredecessor=False, manualLinkToPredecessor=False,
-              inheritAdvices=False, inheritedAdviceUids=[], keep_ftw_labels=False):
+              inheritAdvices=False, inheritedAdviceUids=[], keep_ftw_labels=False,
+              keptAnnexIds=[], keptDecisionAnnexIds=[]):
         '''Clones me in the PloneMeetingFolder of the current user, or
            p_newOwnerId if given (this guy will also become owner of this
            item). If there is a p_cloneEventAction, an event will be included
@@ -5557,7 +5558,10 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
            a manualLink to the predecessor, otherwise, the 'ItemPredecessor' reference is used
            and the link is unbreakable (at least thru the UI).
            If p_inheritAdvices is True, advices will be inherited from predecessor,
-           this also needs p_setCurrentAsPredecessor=True and p_manualLinkToPredecessor=False.'''
+           this also needs p_setCurrentAsPredecessor=True and p_manualLinkToPredecessor=False.
+           When p_copyAnnexes=True, we may give a p_keptAnnexIds, if so, only annexes
+           with those ids are kept, if not, every annexes are kept.
+           Same for p_copyDecisionAnnexes/p_keptDecisionAnnexIds.'''
 
         # check if may clone
         self._mayClone(cloneEventAction)
@@ -5593,12 +5597,16 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
             cloned_to_same_mc, cloned_from_item_template)
 
         # clone
-        newItem = tool.pasteItem(destFolder, copiedData, copyAnnexes=copyAnnexes,
+        newItem = tool.pasteItem(destFolder,
+                                 copiedData,
+                                 copyAnnexes=copyAnnexes,
                                  copyDecisionAnnexes=copyDecisionAnnexes,
                                  newOwnerId=newOwnerId, copyFields=copyFields,
                                  newPortalType=newPortalType,
                                  keepProposingGroup=keepProposingGroup,
-                                 keep_ftw_labels=keep_ftw_labels)
+                                 keep_ftw_labels=keep_ftw_labels,
+                                 keptAnnexIds=keptAnnexIds,
+                                 keptDecisionAnnexIds=keptDecisionAnnexIds)
 
         # special handling for some fields kept when cloned_to_same_mc
         # we check that used values on original item are still useable for cloned item
@@ -5937,32 +5945,6 @@ class MeetingItem(OrderedBaseFolder, BrowserDefaultMixin):
                 else:
                     return brains[0]
         return None
-
-    security.declarePublic('onDuplicate')
-
-    def onDuplicate(self):
-        '''This method is triggered when the users clicks on
-           "duplicate item".'''
-        user = api.user.get_current()
-        newItem = self.clone(newOwnerId=user.id, cloneEventAction=DUPLICATE_EVENT_ACTION)
-        self.plone_utils.addPortalMessage(
-            translate('item_duplicated', domain='PloneMeeting', context=self.REQUEST))
-        return self.REQUEST.RESPONSE.redirect(newItem.absolute_url())
-
-    security.declarePublic('onDuplicateAndKeepLink')
-
-    def onDuplicateAndKeepLink(self):
-        '''This method is triggered when the users clicks on
-           "duplicate item and keep link".'''
-        user = api.user.get_current()
-        newItem = self.clone(newOwnerId=user.id,
-                             cloneEventAction=DUPLICATE_AND_KEEP_LINK_EVENT_ACTION,
-                             setCurrentAsPredecessor=True,
-                             manualLinkToPredecessor=True)
-        plone_utils = api.portal.get_tool('plone_utils')
-        plone_utils.addPortalMessage(
-            translate('item_duplicated_and_link_kept', domain='PloneMeeting', context=self.REQUEST))
-        return self.REQUEST.RESPONSE.redirect(newItem.absolute_url())
 
     security.declarePrivate('manage_beforeDelete')