+++ src/Products/PloneMeeting/skins/plonemeeting_styles/plonemeetingpopups.js
@@ -68,6 +68,18 @@ function contentHistory() {
   });
 }
 
+// the duplicate item action
+function duplicateItem() {
+    jQuery(function($) {
+        // Content history popup
+        $('.apButtonAction_form_duplicate').prepOverlay({
+           subtype: 'ajax',
+           closeselector: '[name="form.buttons.cancel"]',
+        });
+  });
+}
+
+
 // common overlays
 // the content history popup
 function pmCommonOverlays(selector_prefix='') {
@@ -130,6 +142,7 @@ jQuery(document).ready(function($) {
     emergencyChange();
     completenessChange();
     availableMailingLists();
+    duplicateItem();
 
     // inserting methods infos on meeting_view
     tooltipster_helper(selector='.tooltipster-inserting-methods-helper-msg',
@@ -155,6 +168,7 @@ function initializeDashboard(){
     advicesInfos();
     adviceAddEdit();
     contentHistory();
+    duplicateItem();
     pmCommonOverlays();
     listTypeChange();
 }